th_cnd_utils/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "th_cnd_utils"
description = "AI-DO工具库 - 包含各种数据存储和云服务工具的Python库"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "AI-DO开发团队", email = "aido@example.com"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.7"
dependencies = [
    "python-dotenv>=1.0.0",
    "PyMySQL>=1.0.2",
    "redis>=4.5.4",
    "pika>=1.3.1",
    "tablestore>=5.1.0",
    "oss2>=2.17.0",
    "requests>=2.31.0",
    "flask>=2.3.2",
    "flask-cors>=4.0.0",
]
dynamic = ["version"]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=21.0",
    "flake8>=3.8",
]

[project.urls]
Homepage = "https://github.com/your-org/th_cnd_utils"
"Bug Tracker" = "https://github.com/your-org/th_cnd_utils/issues"

[tool.setuptools_scm]
version_scheme = "guess-next-dev"
local_scheme = "dirty-tag"

[tool.setuptools.packages.find]
include = ["th_cnd_utils*"]

[tool.setuptools.package-data]
th_cnd_utils = [
    "examples/*.py",
    "*.md",
    "*.example",
]